servidor/requirements.txt
pyserial==3.5
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask==3.0.0
SQLAlchemy==2.0.23